[WIP] Implement negation into the system
During the tests, elements which are negative in an equation and absent
from hypotheses are true. They must be false.

Solution : Implement, after de system and hypotheses generation, a loop
which will add into Hypotheses all variables which are in Systeme (and
not in Hypotheses) with the state flag "absent"